Don't use u64 in public facing API as iOS doesn't support BigUInt64Array.

+++ i3df/build.rs
@@ -329,7 +329,7 @@ fn main() -> Result<(), Box<dyn Error>> {
         #[derive(Clone, Copy, Debug, Default, Deserialize, Serialize)]
         pub struct Texture {
             // TODO generate from YAML
-            pub file_id: u64,
+            pub file_id: u32,
             pub width: u16,
             pub height: u16,
         }
